Change OS* macro to the COMPILER* macro
Turns out that if you compile something with GCC, even on Windows,
it does not contain Windows' types and intrinsics.
For example, `__int32` should be `int` (probably have to fix it later
for other platforms where int might be 16bit or something, although
it's probably will be redundant).
Another example is `__debugbreak` and `__builtin_trap`.